requirement.txt
google-generativeai
pathlib
flask-socketio
flask